web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "人工评分 Web 界面"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-socketio>=5.6.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pyjwt>=2.10.1",
]
